+++ lib/screens/invoicing/bank_details_screen.dart
@@ -128,7 +128,7 @@ class _BankDetailsScreenState extends State<BankDetailsScreen> {
                     const SizedBox(height: 24),
                     TextFormField(
                       controller: _bankNameController,
-                      keyboardAppearance: Brightness.light,
+                      keyboardAppearance: Theme.of(context).brightness,
                       decoration: const InputDecoration(
                         labelText: 'Bank Name',
                         prefixIcon: Icon(AppIcons.bank),
@@ -138,7 +138,7 @@ class _BankDetailsScreenState extends State<BankDetailsScreen> {
                     const SizedBox(height: 16),
                     TextFormField(
                       controller: _accountNameController,
-                      keyboardAppearance: Brightness.light,
+                      keyboardAppearance: Theme.of(context).brightness,
                       decoration: const InputDecoration(
                         labelText: 'Account Name',
                         prefixIcon: Icon(AppIcons.user),
@@ -148,7 +148,7 @@ class _BankDetailsScreenState extends State<BankDetailsScreen> {
                     const SizedBox(height: 16),
                     TextFormField(
                       controller: _sortCodeController,
-                      keyboardAppearance: Brightness.light,
+                      keyboardAppearance: Theme.of(context).brightness,
                       decoration: const InputDecoration(
                         labelText: 'Sort Code',
                         hintText: 'XX-XX-XX',
@@ -160,7 +160,7 @@ class _BankDetailsScreenState extends State<BankDetailsScreen> {
                     const SizedBox(height: 16),
                     TextFormField(
                       controller: _accountNumberController,
-                      keyboardAppearance: Brightness.light,
+                      keyboardAppearance: Theme.of(context).brightness,
                       decoration: const InputDecoration(
                         labelText: 'Account Number',
                         prefixIcon: Icon(AppIcons.tag),
@@ -171,7 +171,7 @@ class _BankDetailsScreenState extends State<BankDetailsScreen> {
                     const SizedBox(height: 16),
                     TextFormField(
                       controller: _paymentTermsController,
-                      keyboardAppearance: Brightness.light,
+                      keyboardAppearance: Theme.of(context).brightness,
                       decoration: const InputDecoration(
                         labelText: 'Payment Terms',
                         prefixIcon: Icon(AppIcons.clock),