+++ lib/widgets/custom_text_field.dart
@@ -230,7 +230,7 @@ class _CustomTextFieldState extends State<CustomTextField>
         controller: widget.controller,
         focusNode: _focusNode,
         initialValue: widget.initialValue,
-        keyboardAppearance: Brightness.light,
+        keyboardAppearance: Theme.of(context).brightness,
         keyboardType: widget.maxLines > 1 && widget.keyboardType == TextInputType.text
             ? TextInputType.multiline
             : widget.keyboardType == TextInputType.number